Initial Commit
Great news to all people with a Omniksol inverter with wifi module! Now
you can log the data of your inverter and even automatically upload this
data to PVoutput.

For now this workings are fairly simple. This script opens a socket to
the inverter and sends some request data. The response of the inverter
is then parsed by a object found in InverterMsg.py. For uploading to the
mysql database the python mysqldb module is used. And uploading to
PVoutput uses a GET request on th PVoutput API with functions from the
standard urllib module.