Created a Simple Inverter Simulator
This small python script creates a tcp server which acts a if it is a Omnik Inverter. It gets its data from the mysql database the export script uses. The data used is from six hours earlier so development can be done when the sun is down. Also a extra config file is added, in this way the simulator can access data from the real inverter and the script can use a development database.