Added csv output to the LiveStats script
The people from WebSolarLog (http://www.websolarlog.com/) asked if it was possible to add a csv output so they can use this script. So I added a --csv (or --wsl) parameter to the LiveStats script. If that parameter is present the human readable output will be replaced with a csv output.